requirements.txt
google-genai>=1.0.0
psycopg2-binary>=2.9.9
qdrant-client>=1.9.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
